# artifacts/api-server/src/routes/posts.ts
import { Router, type IRouter } from "express";
import { and, asc, count, desc, eq, inArray, sql } from "drizzle-orm";
import { z } from "zod/v4";
import {
  db,
  membersTable,
  postImagesTable,
  postsTable,
  sharesTable,
} from "../db";
import { logger } from "../lib/logger";
import { requireAuth } from "../middleware/auth";
import { HttpError } from "../utils/errors";

const router: IRouter = Router();

const uuidSchema = z.string().uuid();
export const paginationSchema = z.object({
  page: z.coerce.number().int().min(1).default(1),
  limit: z.coerce.number().int().min(1).max(100).default(20),
});
const createPostSchema = z.object({
  caption: z.string().trim().min(1).max(5_000),
  imageUrl: z.string().trim().url().max(2_048).nullable().optional(),
  imageUrls: z
    .array(z.string().trim().url().max(2_048))
    .min(1)
    .max(10)
    .optional(),
});

const postRowSelection = {
  id: postsTable.id,
  authorId: postsTable.authorId,
  caption: postsTable.caption,
  imageUrl: postsTable.imageUrl,
  isFeatured: postsTable.isFeatured,
  featuredAt: postsTable.featuredAt,
  createdAt: postsTable.createdAt,
  updatedAt: postsTable.updatedAt,
  authorName: membersTable.name,
  authorTitle: membersTable.title,
  authorCompany: membersTable.company,
  authorChapter: membersTable.chapter,
  authorProfilePictureUrl: membersTable.profilePictureUrl,
  shareCount: count(sharesTable.id),
};

const postGroupByColumns = [
  postsTable.id,
  postsTable.authorId,
  postsTable.caption,
  postsTable.imageUrl,
  postsTable.isFeatured,
  postsTable.featuredAt,
  postsTable.createdAt,
  postsTable.updatedAt,
  membersTable.name,
  membersTable.title,
  membersTable.company,
  membersTable.chapter,
  membersTable.profilePictureUrl,
] as const;

interface PostRow {
  id: string;
  authorId: string;
  caption: string;
  imageUrl: string | null;
  isFeatured: boolean;
  featuredAt: Date | null;
  createdAt: Date;
  updatedAt: Date;
  authorName: string;
  authorTitle: string;
  authorCompany: string;
  authorChapter: string;
  authorProfilePictureUrl: string | null;
  shareCount: number;
}

export interface SerializedPostImage {
  id: string;
  imageUrl: string;
  position: number;
}

interface SerializedPost {
  id: string;
  caption: string;
  imageUrl: string | null;
  images: SerializedPostImage[];
  status: "featured" | "normal";
  isFeatured: boolean;
  featuredAt: Date | null;
  createdAt: Date;
  updatedAt: Date;
  author: {
    id: string;
    name: string;
    title: string;
    company: string;
    chapter: string;
    profilePictureUrl: string | null;
  };
  shareCount: number;
}

function serializePost(
  row: PostRow,
  postImages: SerializedPostImage[],
): SerializedPost {
  const images =
    postImages.length > 0
      ? postImages
      : row.imageUrl
        ? [{ id: `legacy-${row.id}`, imageUrl: row.imageUrl, position: 0 }]
        : [];

  return {
    id: row.id,
    caption: row.caption,
    imageUrl: images[0]?.imageUrl ?? row.imageUrl,
    images,
    status: row.isFeatured ? "featured" : "normal",
    isFeatured: row.isFeatured,
    featuredAt: row.featuredAt,
    createdAt: row.createdAt,
    updatedAt: row.updatedAt,
    author: {
      id: row.authorId,
      name: row.authorName,
      title: row.authorTitle,
      company: row.authorCompany,
      chapter: row.authorChapter,
      profilePictureUrl: row.authorProfilePictureUrl,
    },
    shareCount: Number(row.shareCount),
  };
}

async function serializePosts(rows: PostRow[]): Promise<SerializedPost[]> {
  if (rows.length === 0) {
    return [];
  }

  const postIds = rows.map((row) => row.id);
  const imageRows = await db
    .select({
      id: postImagesTable.id,
      postId: postImagesTable.postId,
      imageUrl: postImagesTable.imageUrl,
      position: postImagesTable.position,
    })
    .from(postImagesTable)
    .where(inArray(postImagesTable.postId, postIds))
    .orderBy(asc(postImagesTable.position), asc(postImagesTable.id));
  const imagesByPostId = new Map<string, SerializedPostImage[]>();

  for (const image of imageRows) {
    const images = imagesByPostId.get(image.postId) ?? [];
    images.push({
      id: image.id,
      imageUrl: image.imageUrl,
      position: image.position,
    });
    imagesByPostId.set(image.postId, images);
  }

  return rows.map((row) => serializePost(row, imagesByPostId.get(row.id) ?? []));
}

function validateManagedImageUrls(imageUrls: string[]) {
  if (imageUrls.length === 0) {
    return;
  }

  const configuredBaseUrl = process.env.R2_PUBLIC_URL?.trim();
  if (!configuredBaseUrl) {
    throw new HttpError(503, "Image storage is not configured.");
  }

  let baseUrl: URL;
  try {
    baseUrl = new URL(configuredBaseUrl);
  } catch {
    throw new HttpError(500, "Image storage is not configured correctly.");
  }

  const basePath = baseUrl.pathname.endsWith("/")
    ? baseUrl.pathname
    : `${baseUrl.pathname}/`;
  for (const imageUrl of imageUrls) {
    try {
      const parsed = new URL(imageUrl);
      if (
        parsed.protocol !== "https:" ||
        parsed.origin !== baseUrl.origin ||
        !parsed.pathname.startsWith(basePath)
      ) {
        throw new Error("Outside managed storage");
      }
    } catch {
      throw new HttpError(
        400,
        "Posts can only use images uploaded through Aegis Rise.",
      );
    }
  }
}

function resolveImageUrls(input: z.infer<typeof createPostSchema>): string[] {
  const imageUrls = input.imageUrls ?? (input.imageUrl ? [input.imageUrl] : []);
  if (new Set(imageUrls).size !== imageUrls.length) {
    throw new HttpError(400, "Each post image must be unique.");
  }
  validateManagedImageUrls(imageUrls);
  return imageUrls;
}

export function parseId(value: unknown, resource: "post" | "member"): string {
  const parsedId = uuidSchema.safeParse(value);

  if (!parsedId.success) {
    throw new HttpError(400, `A valid ${resource} ID is required.`);
  }

  return parsedId.data;
}

function buildPagination(page: number, limit: number, total: number) {
  const totalPages = total === 0 ? 0 : Math.ceil(total / limit);

  return {
    page,
    limit,
    total,
    totalPages,
    hasNextPage: page < totalPages,
    hasPreviousPage: page > 1 && totalPages > 0,
  };
}

export async function findVisiblePost(postId: string, chapter: string) {
  const [row] = await db
    .select(postRowSelection)
    .from(postsTable)
    .innerJoin(membersTable, eq(postsTable.authorId, membersTable.id))
    .leftJoin(sharesTable, eq(sharesTable.postId, postsTable.id))
    .where(
      and(
        eq(postsTable.id, postId),
        eq(membersTable.status, "active"),
        eq(membersTable.chapter, chapter),
      ),
    )
    .groupBy(...postGroupByColumns)
    .limit(1);

  return row ? (await serializePosts([row]))[0] : undefined;
}

router.post("/posts", requireAuth, async (request, response, next) => {
  try {
    const requestedAutoPost = request.body?.autoPost === true;
    const input = createPostSchema.parse(request.body);
    const imageUrls = resolveImageUrls(input);
    const createdPost = await db.transaction(async (transaction) => {
      const [post] = await transaction
        .insert(postsTable)
        .values({
          authorId: request.user!.id,
          caption: input.caption,
          imageUrl: imageUrls[0] ?? null,
        })
        .returning({ id: postsTable.id });

      if (imageUrls.length > 0) {
        await transaction.insert(postImagesTable).values(
          imageUrls.map((imageUrl, position) => ({
            postId: post.id,
            imageUrl,
            position,
          })),
        );
      }

      return post;
    });

    const post = await findVisiblePost(createdPost.id, request.user!.chapter);
    if (!post) {
      throw new HttpError(500, "The post was created but could not be loaded.");
    }

    logger.info(
      {
        requestId: request.id,
        memberId: request.user!.id,
        postId: post.id,
        requestedAutoPost,
        autoPostHandledBy: "POST /posts/:id/share",
      },
      "Post created; automatic posting is handled by the share endpoint",
    );
    response.status(201).json({ post });
  } catch (error) {
    next(error);
  }
});

router.get("/posts/feed", requireAuth, async (request, response, next) => {
  try {
    const { page, limit } = paginationSchema.parse(request.query);
    const offset = (page - 1) * limit;
    const [rows, totals] = await Promise.all([
      db
        .select(postRowSelection)
        .from(postsTable)
        .innerJoin(membersTable, eq(postsTable.authorId, membersTable.id))
        .leftJoin(sharesTable, eq(sharesTable.postId, postsTable.id))
        .where(
          and(
            eq(membersTable.status, "active"),
            eq(membersTable.chapter, request.user!.chapter),
          ),
        )
        .groupBy(...postGroupByColumns)
        .orderBy(
          sql`${postsTable.isFeatured} DESC`,
          sql`${postsTable.featuredAt} DESC NULLS LAST`,
          desc(postsTable.createdAt),
          desc(postsTable.id),
        )
        .limit(limit)
        .offset(offset),
      db
        .select({ total: count(postsTable.id) })
        .from(postsTable)
        .innerJoin(membersTable, eq(postsTable.authorId, membersTable.id))
        .where(
          and(
            eq(membersTable.status, "active"),
            eq(membersTable.chapter, request.user!.chapter),
          ),
        ),
    ]);

    const total = totals[0]?.total ?? 0;
    response.json({
      posts: await serializePosts(rows),
      pagination: buildPagination(page, limit, total),
    });
  } catch (error) {
    next(error);
  }
});

router.get("/posts/:id", requireAuth, async (request, response, next) => {
  try {
    const postId = parseId(request.params.id, "post");
    const post = await findVisiblePost(postId, request.user!.chapter);

    if (!post) {
      throw new HttpError(404, "Post not found.");
    }

    response.json({ post });
  } catch (error) {
    next(error);
  }
});

router.delete("/posts/:id", requireAuth, async (request, response, next) => {
  try {
    const postId = parseId(request.params.id, "post");
    const [post] = await db
      .select({ authorId: postsTable.authorId })
      .from(postsTable)
      .where(eq(postsTable.id, postId))
      .limit(1);

    if (!post) {
      throw new HttpError(404, "Post not found.");
    }

    if (post.authorId !== request.user!.id) {
      throw new HttpError(403, "You can only delete your own posts.");
    }

    const [deletedPost] = await db
      .delete(postsTable)
      .where(
        and(
          eq(postsTable.id, postId),
          eq(postsTable.authorId, request.user!.id),
        ),
      )
      .returning({ id: postsTable.id });

    if (!deletedPost) {
      throw new HttpError(404, "Post not found.");
    }

    response.status(204).send();
  } catch (error) {
    next(error);
  }
});

router.get(
  "/members/:id/posts",
  requireAuth,
  async (request, response, next) => {
    try {
      const memberId = parseId(request.params.id, "member");
      const { page, limit } = paginationSchema.parse(request.query);
      const offset = (page - 1) * limit;

      const [member] = await db
        .select({
          id: membersTable.id,
          name: membersTable.name,
          title: membersTable.title,
          company: membersTable.company,
          chapter: membersTable.chapter,
          profilePictureUrl: membersTable.profilePictureUrl,
        })
        .from(membersTable)
        .where(
          and(
            eq(membersTable.id, memberId),
            eq(membersTable.status, "active"),
            eq(membersTable.chapter, request.user!.chapter),
          ),
        )
        .limit(1);

      if (!member) {
        throw new HttpError(404, "Member not found.");
      }

      const [rows, totals] = await Promise.all([
        db
          .select(postRowSelection)
          .from(postsTable)
          .innerJoin(membersTable, eq(postsTable.authorId, membersTable.id))
          .leftJoin(sharesTable, eq(sharesTable.postId, postsTable.id))
          .where(
            and(
              eq(postsTable.authorId, memberId),
              eq(membersTable.status, "active"),
              eq(membersTable.chapter, request.user!.chapter),
            ),
          )
          .groupBy(...postGroupByColumns)
          .orderBy(desc(postsTable.createdAt), desc(postsTable.id))
          .limit(limit)
          .offset(offset),
        db
          .select({ total: count(postsTable.id) })
          .from(postsTable)
        .innerJoin(membersTable, eq(postsTable.authorId, membersTable.id))
        .where(
          and(
            eq(postsTable.authorId, memberId),
            eq(membersTable.status, "active"),
            eq(membersTable.chapter, request.user!.chapter),
          ),
        ),
      ]);

      const total = totals[0]?.total ?? 0;
      response.json({
        member,
        posts: await serializePosts(rows),
        pagination: buildPagination(page, limit, total),
      });
    } catch (error) {
      next(error);
    }
  },
);

export default router;