# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Aegis Rise member, social feed, and chapter administration API.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /members/{memberId}/posts:
    get:
      operationId: listMemberPosts
      tags: [members]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/MemberId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Member posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberPostsResult"
  /posts:
    post:
      operationId: createPost
      tags: [posts]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
  /posts/feed:
    get:
      operationId: getFeed
      tags: [posts]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Chapter feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedResult"
  /posts/{postId}:
    get:
      operationId: getPost
      tags: [posts]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PostId"
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
    delete:
      operationId: deleteOwnPost
      tags: [posts]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PostId"
      responses:
        "204":
          description: Post deleted
  /posts/{postId}/share:
    post:
      operationId: sharePost
      tags: [shares]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PostId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShareInput"
      responses:
        "201":
          description: Share tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareResult"
...
  /admin/posts:
    get:
      operationId: listAdminPosts
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Chapter"
        - name: featured
          in: query
          schema: { type: boolean }
        - name: authorId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Moderatable posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPostList"
...
  /admin/posts/{postId}:
    delete:
      operationId: deleteAdminPost
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PostId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModerationReason"
      responses:
        "200":
          description: Post deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
  /admin/analytics/posts:
    get:
      operationId: getPostAnalytics
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Chapter"
      responses:
        "200":
          description: Leading posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostAnalyticsResult"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Chapter:
      name: chapter
      in: query
      required: false
      schema: { type: string }
    MemberId:
      name: memberId
      in: path
      required: true
      schema: { type: string, format: uuid }
    PostId:
      name: postId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    Message:
      type: object
      required: [message]
      properties: { message: { type: string } }
    Pagination:
      type: object
      required: [page, limit, total, totalPages, hasNextPage, hasPreviousPage]
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
        hasNextPage: { type: boolean }
        hasPreviousPage: { type: boolean }
...
    PostAuthor:
      type: object
      required: [id, name, title, company, chapter]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        title: { type: string }
        company: { type: string }
        chapter: { type: string }
        profilePictureUrl: { type: [string, "null"], format: uri }
    Post:
      type: object
      required: [id, caption, images, status, isFeatured, createdAt, updatedAt, author, shareCount]
      properties:
        id: { type: string, format: uuid }
        caption: { type: string }
        imageUrl: { type: [string, "null"], format: uri }
        images:
          type: array
          maxItems: 10
          items: { $ref: "#/components/schemas/PostImage" }
        status: { type: string, enum: [featured, normal] }
        isFeatured: { type: boolean }
        featuredAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        author: { $ref: "#/components/schemas/PostAuthor" }
        shareCount: { type: integer }
    PostImage:
      type: object
      required: [id, imageUrl, position]
      properties:
        id: { type: string, format: uuid }
        imageUrl: { type: string, format: uri }
        position: { type: integer, minimum: 0 }
    PostInput:
      type: object
      required: [caption]
      properties:
        caption: { type: string, minLength: 1, maxLength: 5000 }
        imageUrl: { type: [string, "null"], format: uri }
        imageUrls:
          type: array
          minItems: 1
          maxItems: 10
          items: { type: string, format: uri }
    PostResponse:
      type: object
      required: [post]
      properties: { post: { $ref: "#/components/schemas/Post" } }
    FeedResult:
      type: object
      required: [posts, pagination]
      properties:
        posts:
          type: array
          items: { $ref: "#/components/schemas/Post" }
        pagination: { $ref: "#/components/schemas/Pagination" }
    MemberPostsResult:
      type: object
      required: [member, posts, pagination]
      properties:
        member: { $ref: "#/components/schemas/PostAuthor" }
        posts:
          type: array
          items: { $ref: "#/components/schemas/Post" }
        pagination: { $ref: "#/components/schemas/Pagination" }
    ShareInput:
      type: object
      required: [platform]
      properties:
        platform: { type: string, enum: [LinkedIn, Instagram, Facebook, TikTok, Direct Link] }
        autoPost: { type: boolean, description: Attempt provider posting to the member's selected connected accounts. }
    ShareResult:
      type: object
      required: [shareCount, message, autoPosted]
      properties:
        shareCount: { type: integer }
        message: { type: string }
        autoPosted:
          type: object
          additionalProperties: { $ref: "#/components/schemas/SocialPostOutcome" }
...
    SocialPostOutcome:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [success, error] }
        externalUrl: { type: string, format: uri }
        error: { type: string }
...
    AdminPost:
      type: object
      required: [postId, author, caption, images, shares, createdAt, updatedAt, status]
      properties:
        postId: { type: string, format: uuid }
        author: { $ref: "#/components/schemas/PostAuthor" }
        caption: { type: string }
        imageUrl: { type: [string, "null"], format: uri }
        images:
          type: array
          maxItems: 10
          items: { $ref: "#/components/schemas/PostImage" }
        shares: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        featuredAt: { type: [string, "null"], format: date-time }
        status: { type: string, enum: [featured, normal] }
    AdminPostList:
      type: object
      required: [posts, pagination]
      properties:
        posts:
          type: array
          items: { $ref: "#/components/schemas/AdminPost" }
        pagination: { $ref: "#/components/schemas/Pagination" }
...
    ModerationReason:
      type: object
      properties: { reason: { type: string, maxLength: 1000 } }
...
    PostAnalyticsResult:
      type: object
      required: [posts]
      properties:
        posts:
          type: array
          items:
            type: object
            required: [id, caption, author, shares, createdAt]
            properties:
              id: { type: string, format: uuid }
              caption: { type: string }
              author: { type: string }
              shares: { type: integer }
              image: { type: [string, "null"], format: uri }
              createdAt: { type: string, format: date-time }
...