# lib/api-zod/src/generated/types/post.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Aegis Rise member, social feed, and chapter administration API.
 * OpenAPI spec version: 0.1.0
 */
import type { PostAuthor } from './postAuthor';
import type { PostImage } from './postImage';
import type { PostStatus } from './postStatus';

export interface Post {
  id: string;
  caption: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @maxItems 10 */
  images: PostImage[];
  status: PostStatus;
  isFeatured: boolean;
  /** @nullable */
  featuredAt?: Date | null;
  createdAt: Date;
  updatedAt: Date;
  author: PostAuthor;
  shareCount: number;
}