# lib/api-zod/src/generated/types/postInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Aegis Rise member, social feed, and chapter administration API.
 * OpenAPI spec version: 0.1.0
 */

export interface PostInput {
  /**
     * @minLength 1
     * @maxLength 5000
     */
  caption: string;
  /** @nullable */
  imageUrl?: string | null;
  /**
     * @minItems 1
     * @maxItems 10
     */
  imageUrls?: string[];
}